feat: add Delete Option

+++ server/app/components/option/optionController.js
@@ -52,7 +52,6 @@ exports.updateOption = async (req, res, next) => {
   for (let x in optionList) {
     const option = await optionService.findById(optionList[x].id);
     if (!option || !await slide.hasOption(option)) {
-      console.log(await slide.hasOption(option));
       return res.status(404).send({ message: "Option not found" });
     }
     await optionService.update(optionList[x].id, optionList[x].content, optionList[x].isCorrect);
@@ -63,7 +62,15 @@ exports.updateOption = async (req, res, next) => {
 
 exports.deleteOption = async (req, res, next) => {
   const slideId = req.params["slideId"];
+  const slide = await slideService.findById(slideId);
+  if (!slide) {
+    return res.status(404).send({ message: "Slide not found" });
+  }
   const optionId = req.params["optionId"];
+  const option = await optionService.findById(optionId);
+  if (!option || !await slide.hasOption(option)) {
+    return res.status(404).send({ message: "Option not found" });
+  }
   await slideService.delete(slideId);
   return res.status(200).send({ message: "Delete successfully!" });
 };
